% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_dataset_metadata}
\alias{add_dataset_metadata}
\title{Add metadata to a dataset as a whole}
\usage{
add_dataset_metadata(data, metadata_list)
}
\arguments{
\item{data}{Data frame}
}
\description{
Add metadata to a dataset as a whole
}
