% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_variable_metadata}
\alias{add_variable_metadata}
\title{Add metadata to columns in a data frame}
\usage{
add_variable_metadata(data, fields, try_to_factorize = "never")
}
\arguments{
\item{data}{Data frame}

\item{fields}{Named list of metadata fields to add to data}

\item{try_to_factorize}{Whether or not to coerce to factor}
}
\value{
variable
}
\description{
Add metadata to columns in a data frame
}
