% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{attr_from_field}
\alias{attr_from_field}
\title{Set an object's attribute from a list object}
\usage{
attr_from_field(variable, attribute, fields, try_to_factorize = NULL)
}
\arguments{
\item{variable}{Object to be modified}

\item{attribute}{Character string of attribute to apply to variable}

\item{fields}{Named list of fields from which to get attribute}

\item{try_to_factorize}{Whether or not to coerce to factor}
}
\value{
variable, with additional attribute added
}
\description{
Note that the "type" attribute is a special case.
}
