% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colclass_dict.R
\docType{data}
\name{colclass_dict}
\alias{colclass_dict}
\title{Dictionary of column classes for reading data}
\format{An object of class \code{character} of length 7.}
\usage{
colclass_dict
}
\description{
Dictionary of column classes for reading data
}
\keyword{datasets}
