% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csvy.R
\name{field_from_attr}
\alias{field_from_attr}
\title{Set field from attributes list}
\usage{
field_from_attr(field, attribute, values)
}
\arguments{
\item{field}{List to which to add attributes}

\item{attribute}{Character name of attribute to add}

\item{values}{List of attributes, returned by \code{attributes(x)}}
}
\value{
Modified field list with given attribute set to given value
}
\description{
Set field from attributes list
}
