% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yaml_header.R
\name{get_yaml_header}
\alias{get_yaml_header}
\title{Retrieve YAML header from file}
\usage{
get_yaml_header(file, yaml_rxp = "^#*---[[:space:]]*$", verbose = TRUE)
}
\arguments{
\item{file}{A character string or R connection specifying a file.}

\item{yaml_rxp}{Regular expression for parsing YAML header}

\item{verbose}{Logical. If \code{TRUE}, print warning if no header found.}
}
\value{
Character vector of lines containing YAML header, or `NULL` if no YAML header found.
}
\description{
Note that this assumes only one Yaml header, starting on the first line of the file.
}
