% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(file)
}
\arguments{
\item{file}{full path of file from which to read the metadata.}
}
\value{
the metadata as a list
}
\description{
Read csvy metadata from an external \code{.yml/.yaml} or \code{.json} file
}
