% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{type_from_field}
\alias{type_from_field}
\title{Set an object's type based on a list of fields}
\usage{
type_from_field(variable, fields, try_to_factorize)
}
\arguments{
\item{fields}{Named list of metadata fields to add to data}

\item{try_to_factorize}{Whether or not to coerce to factor}
}
\description{
Set an object's type based on a list of fields
}
