% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_dataset_metadata}
\alias{add_dataset_metadata}
\title{Add metadata to a dataset as a whole}
\usage{
add_dataset_metadata(data, metadata_list)
}
\arguments{
\item{data}{Data frame whose columns should be modified}

\item{metadata_list}{List of metadata fields to add to data}
}
\value{
data Modified data object with attributes added to columns
}
\description{
Add metadata to a dataset as a whole
}
