% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csvy.R
\name{read_csvy}
\alias{read_csvy}
\title{Import CSVY data}
\usage{
read_csvy(file, sep = "auto", sep2 = "auto", header = "auto",
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{file}{A character string or R connection specifying a file.}

\item{sep}{A character string specifying a between-field separator. Passed to \code{\link[data.table]{fread}}.}

\item{sep2}{A character string specifying a within-field separator. Passed to \code{\link[data.table]{fread}}.}

\item{header}{A character string or logical specifying whether the file contains a header row of column names (below the YAML frontmatter). Passed to \code{\link[data.table]{fread}}.}

\item{stringsAsFactors}{A logical specifying whether to treat character columns as factors. Passed to \code{\link[data.table]{fread}}.}

\item{\dots}{Additional arguments passed to \code{\link[data.table]{fread}}.}
}
\description{
Import CSVY data as a data.frame
}
\seealso{
\code{\link{write_csvy}}
}

